Version 1.0.1
Add downloader built-in

Project can be the normal gradle plugin but can also be a dependency. The depedency part allows you to easily give the meta-inf/dependencies.json back and it will download then load those jars into the classloader.

Added basic json parsing library. Nothing to crazy but allows use to parse and create json without a library so it can be a standalone jar still.